+++ usr.bin/dc/extern.h
@@ -47,7 +47,7 @@ void		stack_init(struct stack *);
 void		stack_free_value(struct value *);
 struct value	*stack_dup_value(const struct value *, struct value *);
 void		stack_swap(struct stack *);
-int		stack_size(const struct stack *);
+size_t		stack_size(const struct stack *);
 void		stack_dup(struct stack *);
 void		stack_pushnumber(struct stack *, struct number *);
 void		stack_pushstring(struct stack *stack, char *);
